Expose model transaction status
